requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
openpyxl==3.1.5
pydantic==2.9.2
pulp==2.9.0
python-multipart>=0.0.9
msal>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT[cryptography]>=2.8.0